Adding routes update delete user

+++ app/controllers/api/v1/userController.js
@@ -40,6 +40,11 @@ function createToken(payload) {
 
 module.exports = {
     list(req, res) {
+        if (req.user.role === "member") {
+            res.status(401).json({ message: `Role ${req.user.role} is not allowed here` })
+            return
+        }
+
         userService.list()
             .then(({ data }) => {
                 res.status(200).json({
@@ -56,6 +61,11 @@ module.exports = {
     },
 
     detail(req, res) {
+        if (req.user.role === "member") {
+            res.status(401).json({ message: `Role ${req.user.role} is not allowed here` })
+            return
+        }
+
         userService.get(req.params.id)
             .then((user) => {
                 res.status(200).json({
@@ -71,7 +81,25 @@ module.exports = {
             })
     },
 
-    destroy(req, res) {
+    async destroy(req, res) {
+        if (req.user.role !== "superadmin") {
+            res.status(401).json({ message: `Role ${req.user.role} is not allowed here` })
+            return
+        }
+
+        const user = await userService.get(req.params.id)
+
+        const userImage = user.dataValues.image
+
+        const imageID = userImage.split("/").pop().split(".")[0]
+
+        const deleteImage = await cloudinary.uploader.destroy(`user/${imageID}`)
+
+        if (!deleteImage) {
+            res.status(422).json({ message: "Error when deleting file" })
+            return
+        }
+
         userService.delete(req.params.id)
             .then(() => {
                 res.status(204).end()
@@ -82,6 +110,66 @@ module.exports = {
                     errors: err.message
                 })
             })
+    },
+
+    async update(req, res) {
+        if (req.user.role !== "superadmin") {
+            res.status(401).json({ message: `Role ${req.user.role} is not allowed here` })
+            return
+        }
+        
+        if (req.file == null) {
+            userService.update(req.params.id, req.body)
+                .then(() => {
+                    res.status(200).json({ message: "Update successfully" })
+                })
+                .catch((err) => {
+                    res.status(422).json({
+                        status: "FAIL",
+                        errors: err.message
+                    })
+                })
+        } else {
+            const user = await userService.get(req.params.id)
+
+            const userImage = user.dataValues.image
+
+            if (userImage !== null) {
+                const imageID = userImage.split("/").pop().split(".")[0]
+
+                await cloudinary.uploader.destroy(`user/${imageID}`)
+            }
+
+            const fileBase64 = req.file.buffer.toString("base64")
+            const file = `data:${req.file.mimetype};base64,${fileBase64}`
+
+            cloudinary.uploader.upload(file, { folder: 'user' }, function (err, result) {
+                if (!!err) {
+                    res.status(400).json({
+                        status: "UPLOAD FAIL",
+                        errors: err.message
+                    })
+                    return
+                }
+
+                const body = {
+                    ...req.body,
+                    image: result.url
+                }
+
+                userService.update(req.params.id, body)
+                    .then(() => {
+                        res.status(200).json({ message: "Update successfully" })
+                    })
+                    .catch((err) => {
+                        res.status(400).json({
+                            status: "FAIL",
+                            errors: err.message
+                        })
+                    })
+            })
+
+        }
     },
 
     async whoIsLogin(req, res, next) {
@@ -136,6 +224,8 @@ module.exports = {
         const email = req.body.email.toLowerCase()
         const encryptedPassword = await encyptedPassword(req.body.password)
 
+        const rolePicker = typeof req.user === 'undefined' ? "member" : req.user.role === "superadmin" ? "admin" : "member"
+
         const isEmailExists = await User.findOne({ where: { email } })
 
         if (isEmailExists) {
@@ -147,6 +237,7 @@ module.exports = {
             const body = {
                 ...req.body,
                 email: email,
+                role: rolePicker,
                 password: encryptedPassword
             }
 
@@ -179,6 +270,7 @@ module.exports = {
                 const body = {
                     ...req.body,
                     email: email,
+                    role: rolePicker,
                     password: encryptedPassword,
                     image: result.url
                 }
